Transformation of Effect Sizes

Converts between different measures of effect size (i.e., Cohen's d, log odds ratio, Pearson correlation r, and Fisher's z).

transform_es(y, SE, from, to)

Arguments

y
estimate of the effect size (can be vectorized).

SE
optional: standard error of the effect-size estimate. Must have the same length as y.

from
type of effect-size measure provided by the argument y. Supported effect sizes are Cohen's d ("d"), Fisher's z-transformed correlation ("z"), Pearson's correlation ("r"), or the log odds ratio ("logOR").

to
which type of effect size should be returned (see from).

Value

If SE is missing, a vector of transformed effect sizes. Otherwise, a matrix with two columns including effect sizes and standard errors.

Details

The following chain of transformations is adopted from Borenstein et al. (2009): logOR <--> d <--> r <--> z. The conversion from "d" to "r" assumes equal sample sizes per condition (n1=n2).

Note that in in a Bayesian meta-analysis, the prior distributions need to be adapted to the type of effect size. The function meta_default provides modified default prior distributions for different effect size measures which are approximately transformation-invariant (but results may still differ depending on which type of effect size is used for analysis).

References

Borenstein, M., Hedges, L. V., Higgins, J. P. T., & Rothstein, H. R. (2009). Converting among effect sizes. In Introduction to Meta-Analysis (pp. 45–49). John Wiley & Sons, Ltd. doi: 10.1002/9780470743386.ch7

See also

Examples

# transform a single value of Cohen's
transform_es(y = 0.50, SE = 0.20, from = "d", to = "logOR")

#>          logOR        SE
#> [1,] 0.9068997 0.3627599

# towels data set:
transform_es(y = towels$logOR, SE = towels$SE, from = "logOR", to = "d")

#>                d         SE
#> [1,]  0.20983674 0.10899227
#> [2,]  0.16812095 0.07504441
#> [3,]  0.11331896 0.10563023
#> [4,]  0.13829794 0.09376833
#> [5,]  0.15860744 0.45415964
#> [6,] -0.06700986 0.13675933
#> [7,] -0.80379564 0.41910909