Data Set: Reuse of Towels in Hotels

Set of studies that investigated whether people reuse towels in hotels more often if they are provided with a descriptive norm (Scheibehenne, Jamil, & Wagenmakers, 2016).

towels

Format

A data frame with three variables:

study

Authors of original study (see Scheibehenne et. al, 2016)

logOR

Measure of effect size: log-odds ratio of towel reuse (descriptive-social-norm vs. control)

SE

Measure of precision: standard error of log-odds ratio per study

Details

Two groups of hotel guests received different messages that encouraged them to reuse their towels. One message simply informed the guests about the benefits of environmental protection (the control condition), and the other message indicated that the majority of guests actually reused their towels in the past (the descriptive-social-norm condition). The results suggested that the latter message facilitated towel reuse.

References

Scheibehenne, B., Jamil, T., & Wagenmakers, E.-J. (2016). Bayesian Evidence Synthesis Can Reconcile Seemingly Inconsistent Results: The Case of Hotel Towel Reuse. Psychological Science, 27, 1043–1046. doi:10.1177/0956797616644081

Examples

data(towels)
head(towels)
#>                                                 study      logOR        SE
#> 3  Goldstein, Cialdini, & Griskevicius (2008), Exp. 1  0.3806017 0.1976901
#> 4 Goldstein, Cialdini, & Griskevicius  (2008), Exp. 2  0.3049377 0.1361155
#> 6          Schultz, Khazian, & Zaleski (2008), Exp. 2  0.2055379 0.1915920
#> 7          Schultz, Khazian, & Zaleski (2008), Exp. 3  0.2508447 0.1700769
#> 5                  Mair & Bergin-Seers (2010), Exp. 1  0.2876821 0.8237545
#> 1                    Bohner & Schluter (2014), Exp. 1 -0.1215424 0.2480540